requirements.txt
PyMuPDF==1.23.26
pandas==2.1.4
scikit-learn==1.3.2
xgboost==1.7.6
numpy==1.26.4
PyYAML==6.0.1
